httpcaddyfile: fix placeholder shorthands in named routes (#5791)
