+++ .github/workflows/lint.yml
@@ -56,3 +56,5 @@ jobs:
     steps:
       - name: govulncheck
         uses: golang/govulncheck-action@v1
+        with:
+          check-latest: true
